backend/pyproject.toml
[project]
name = "fastapi-chat-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Yousuf Traders",email = "yousufraders777@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.124.4,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "google-generativeai (>=0.8.6,<0.9.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "loguru (>=0.7.2,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "aiosqlite (>=0.22.0,<0.23.0)",
    "httpx"
]
